ensure all changes on list of gc objects are done in VM thread
fixes memleaks and crash on exit
in "FormulaRacer2012" and "Super Mario 63"

+++ src/scripting/toplevel/ASString.cpp
@@ -758,7 +758,7 @@ string ASString::toDebugString() const
 		ret = std::string("\"") + std::string(data) + "\"";
 #ifndef NDEBUG
 	char buf[300];
-	sprintf(buf,"(%p/%d/%d/%d%s) ",this,this->getRefCount(),this->storedmembercount,this->getConstant(),this->isConstructed()?"":" not constructed");
+	sprintf(buf,"(%p/%i/%i/%i%s) ",this,this->getRefCount(),(int32_t)this->storedmembercount,this->getConstant(),this->isConstructed()?"":" not constructed");
 	ret += buf;
 #endif
 	return ret;
